requirements.txt
biopython
numpy
pandas
scikit-learn
matplotlib
tensorflow
transformers